+++ tket/test/src/test_GreedyPauli.cpp
@@ -27,6 +27,7 @@
 #include "tket/Predicates/CompilerPass.hpp"
 #include "tket/Predicates/PassGenerators.hpp"
 #include "tket/Predicates/PassLibrary.hpp"
+#include "tket/Transformations/BasicOptimisation.hpp"
 #include "tket/Transformations/GreedyPauliOptimisation.hpp"
 #include "tket/Utils/Expression.hpp"
 
@@ -368,6 +369,23 @@ SCENARIO("Complete synthesis") {
     REQUIRE(Transforms::greedy_pauli_optimisation().apply(circ));
     REQUIRE(circ == d);
   }
+
+  GIVEN("Circuit with conditional gates 3") {
+    Circuit c(2, 1);
+    c.add_op<unsigned>(OpType::Rx, 1.0246, {0});
+    c.add_conditional_gate<unsigned>(OpType::Ry, {-0.25}, {1}, {0}, 0);
+    c.add_conditional_gate<unsigned>(OpType::CZ, {}, {0, 1}, {0}, 0);
+    REQUIRE(Transforms::greedy_pauli_optimisation().apply(c));
+
+    Circuit comp(2, 1);
+    comp.add_op<unsigned>(OpType::Rx, 1.0246, {0});
+    comp.add_conditional_gate<unsigned>(OpType::Ry, {-0.25}, {1}, {0}, 0);
+    comp.add_conditional_gate<unsigned>(OpType::CZ, {}, {1, 0}, {0}, 0);
+    // Previously Conditional Ry was being commuted ahead of Rz and then CZ
+    // merged in without checking commuting with Rx Returning the same circuit
+    // is expected, except the CZ gate is "flipped"
+    REQUIRE(comp == c);
+  }
   GIVEN("Circuit with conditional gates and measures") {
     Circuit circ(2, 2);
     Op_ptr cond1 = std::make_shared<Conditional>(
@@ -405,6 +423,7 @@ SCENARIO("Complete synthesis") {
     // should all be canceled
     Circuit d(2, 2);
     REQUIRE(Transforms::greedy_pauli_optimisation().apply(circ));
+    Transforms::remove_redundancies().apply(circ);
     REQUIRE(circ == d);
   }
   GIVEN("Circuit with classical gates") {
@@ -677,7 +696,8 @@ SCENARIO("Complete synthesis") {
             .apply(d));
     REQUIRE(test_unitary_comparison(circ, d, true));
     // if the first XY was implemented using a ZZPhase
-    // then 2 TQEs is needed to conjugate the remaining two strings to weight 2
+    // then 2 TQEs is needed to conjugate the remaining two strings to
+    // weight 2
     // hence 5 2-qubit gates in total.
     REQUIRE(d.count_n_qubit_gates(2) == 4);
   }